requirements.txt
alembic==1.6.5
fastapi==0.65.2
pydantic==1.8.2
PyMySQL==1.0.2
SQLAlchemy==1.4.17
SQLAlchemy-Utils==0.37.7
uvicorn==0.14.0
pytest~=6.2.4
pytest-cov==2.8.1
pyhumps==3.8.0
starlette==0.14.2
starlette-context==0.3.3
python-dotenv==0.17.1
requests==2.25.1
prometheus-client==0.7.1
opentelemetry-exporter-jaeger==1.7.1
opentelemetry-instrumentation-fastapi==0.26b1
opentelemetry-instrumentation-requests==0.26b1
opentelemetry-instrumentation-sqlalchemy==0.26b1
opentelemetry-sdk==1.7.1
opentelemetry-api==1.7.1
numpy==1.24.2
pandas==2.0.0
python_multipart==0.0.6
openpyxl==3.0.7
cachetools~=4.2.2